requirements.txt
python-dotenv==1.2.1
Requests==2.32.5
selenium==4.38.0
telegram_notifier==0.3